+++ src/main/java/sptech/school/dto/itensNaMovimentacao/ItensNaMovimentacaoResponseDto.java
@@ -3,11 +3,13 @@ package sptech.school.dto.itensNaMovimentacao;
 import sptech.school.dto.item.ItemResponseDto;
 import sptech.school.dto.movimentacaoEstoque.MovimentacaoEstoqueResponseDto;
 
+import java.math.BigDecimal;
+
 public record ItensNaMovimentacaoResponseDto(
         Integer id,
         MovimentacaoEstoqueResponseDto movimentacaoEstoque,
         ItemResponseDto item,
         Integer qtd,
-        Double precoUnitario
+        BigDecimal precoUnitario
 ) {
 }
